Restructure types and exports
