+++ src/PgManyToManyRelationInflectionPlugin.ts
@@ -1,61 +1,8 @@
-import {
-  PgSource,
-  PgSourceRelation,
-  PgTypeCodec,
-  PgTypeColumn,
-} from "@dataplan/pg";
 import type {} from "graphile-config";
 import type {} from "postgraphile";
-import { PgTableSource } from "./interfaces";
+import { PgManyToManyRelationDetails } from "./interfaces";
 
 const version = require("../package.json").version;
-
-type InflectionColumn = {
-  columnName: string;
-  column: PgTypeColumn;
-  codec: PgTypeCodec<any, any, any>;
-};
-
-export interface PgManyToManyRelationDetails {
-  leftTable: PgTableSource;
-  leftRelationName: string;
-  junctionTable: PgTableSource;
-  rightRelationName: string;
-  rightTable: PgTableSource;
-  allowsMultipleEdgesToNode: boolean;
-}
-
-export interface PgManyToManyRelationDetailsWithExtras
-  extends PgManyToManyRelationDetails {
-  leftTableTypeName: string;
-}
-
-declare global {
-  namespace GraphileBuild {
-    interface Inflection {
-      manyToManyRelationByKeys(
-        this: Inflection,
-        details: PgManyToManyRelationDetails
-      ): string;
-      manyToManyRelationByKeysSimple(
-        this: Inflection,
-        details: PgManyToManyRelationDetails
-      ): string;
-      manyToManyRelationEdge(
-        this: Inflection,
-        details: PgManyToManyRelationDetailsWithExtras
-      ): string;
-      manyToManyRelationConnection(
-        this: Inflection,
-        details: PgManyToManyRelationDetailsWithExtras
-      ): string;
-      manyToManyRelationSubqueryName(
-        this: Inflection,
-        details: PgManyToManyRelationDetailsWithExtras
-      ): string;
-    }
-  }
-}
 
 const processDetails = (details: PgManyToManyRelationDetails) => {
   const {
@@ -63,7 +10,7 @@ const processDetails = (details: PgManyToManyRelationDetails) => {
     leftRelationName,
     junctionTable,
     rightRelationName,
-    rightTable,
+    // rightTable,
   } = details;
   const junctionRightRelation = junctionTable.getRelation(rightRelationName);
   const columnInflectionDataFromJunction = (columnName: string) => {
@@ -92,7 +39,7 @@ export const PgManyToManyRelationInflectionPlugin: GraphileConfig.Plugin = {
 
   inflection: {
     add: {
-      manyToManyRelationByKeys(preset, details) {
+      manyToManyRelationByKeys(_preset, details) {
         const {
           junctionRightKeyAttributes,
           junctionLeftKeyAttributes,
@@ -117,7 +64,7 @@ export const PgManyToManyRelationInflectionPlugin: GraphileConfig.Plugin = {
             .join("-and-")}`
         );
       },
-      manyToManyRelationByKeysSimple(preset, details) {
+      manyToManyRelationByKeysSimple(_preset, details) {
         const {
           junctionRightKeyAttributes,
           junctionLeftKeyAttributes,
@@ -143,20 +90,20 @@ export const PgManyToManyRelationInflectionPlugin: GraphileConfig.Plugin = {
             .join("-and-")}-list`
         );
       },
-      manyToManyRelationEdge(preset, details) {
+      manyToManyRelationEdge(_preset, details) {
         const relationName = this.manyToManyRelationByKeys(details);
         return this.upperCamelCase(
           `${details.leftTableTypeName}-${relationName}-many-to-many-edge`
         );
       },
-      manyToManyRelationConnection(preset, details) {
+      manyToManyRelationConnection(_preset, details) {
         const relationName = this.manyToManyRelationByKeys(details);
         return this.upperCamelCase(
           `${details.leftTableTypeName}-${relationName}-many-to-many-connection`
         );
       },
       /* eslint-disable no-unused-vars */
-      manyToManyRelationSubqueryName(preset, details) {
+      manyToManyRelationSubqueryName(_preset, details) {
         /* eslint-enable no-unused-vars */
         return `many-to-many-subquery-by-${this._singularizedCodecName(
           details.junctionTable.codec
